requirements.txt
accelerate==1.6.0
bitsandbytes==0.45.5
deepspeed==0.16.5
flash-attn==2.7.2.post1
huggingface-hub==0.30.2
hydra-core==1.3.2
liger_kernel==0.5.8
numpy==1.26.0
omegaconf==2.3.0
pandas==1.5.0
peft==0.15.1
pyarrow==17.0.0
python-dotenv==1.0.1
safetensors==0.5.3
scikit-learn==1.5.2
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
transformers==4.51.3
triton==3.1.0
trl==0.16.1
wandb==0.18.3